chore(ui) Remove remaining types from types/index
This leaves types/index.tsx as a simple entrypoint that we can import
from for better ergonomics.
